[task] Added termite config
Termite terminal will now be used to replace urxvt as urxvt does not
not work around unicode font issues with variaous fonts. This results
in undisplayed characters which is of no benefit.

Temrite has also be configured to use the solarize theme as its easier
on the eyes. This theme will be propagated to tmux add vim